+++ lib/dataset/variable_reduction.cpp
@@ -9,37 +9,68 @@
 #include "scipp/variable/creation.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/special_values.h"
-#include "scipp/variable/transform.h"
 #include "scipp/variable/util.h"
 
 #include "dataset_operations_common.h"
 
 namespace scipp::dataset {
 
-Variable sum(const Variable &var, const Dim dim, const Masks &masks) {
-  if (const auto mask_union = irreducible_mask(masks, dim);
-      mask_union.is_valid()) {
-    return sum(where(mask_union, zero_like(var), var), dim);
-  }
-  return sum(var, dim);
+namespace {
+// Uses variable::special like but constructs only a scalar
+// instead of a full-sized array.
+Variable scalar_special_like(const Variable &prototype, const FillValue fill) {
+  return special_like(zero_like(prototype), fill);
 }
 
-Variable &sum(const Variable &var, const Dim dim, const Masks &masks,
-              // cppcheck-suppress constParameter  # intentional out param
-              Variable &out) {
+template <class Op>
+Variable reduce_impl(const Variable &var, const Dim dim, const Masks &masks,
+                     const FillValue fill, const Op &op) {
   if (const auto mask_union = irreducible_mask(masks, dim);
       mask_union.is_valid()) {
-    return sum(where(mask_union, zero_like(var), var), dim, out);
+    return op(where(mask_union, scalar_special_like(var, fill), var), dim);
   }
-  return sum(var, dim, out);
+  return op(var, dim);
+}
+} // namespace
+
+Variable sum(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::Default,
+                     [](auto &&... args) { return sum(args...); });
 }
 
 Variable nansum(const Variable &var, const Dim dim, const Masks &masks) {
-  if (const auto mask_union = irreducible_mask(masks, dim);
-      mask_union.is_valid()) {
-    return nansum(where(mask_union, zero_like(var), var), dim);
-  }
-  return nansum(var, dim);
+  return reduce_impl(var, dim, masks, FillValue::Default,
+                     [](auto &&... args) { return nansum(args...); });
+}
+
+Variable max(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::Lowest,
+                     [](auto &&... args) { return max(args...); });
+}
+
+Variable nanmax(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::Lowest,
+                     [](auto &&... args) { return nanmax(args...); });
+}
+
+Variable min(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::Max,
+                     [](auto &&... args) { return min(args...); });
+}
+
+Variable nanmin(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::Max,
+                     [](auto &&... args) { return nanmin(args...); });
+}
+
+Variable all(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::True,
+                     [](auto &&... args) { return all(args...); });
+}
+
+Variable any(const Variable &var, const Dim dim, const Masks &masks) {
+  return reduce_impl(var, dim, masks, FillValue::False,
+                     [](auto &&... args) { return any(args...); });
 }
 
 Variable mean(const Variable &var, const Dim dim, const Masks &masks) {