+++ src/scipp/core/reduction.py
@@ -6,14 +6,10 @@ from __future__ import annotations
 from typing import Optional
 
 from .._scipp import core as _cpp
-from ._cpp_wrapper_util import call_func as _call_cpp_func
 from ..typing import VariableLikeType
 
 
-def mean(x: VariableLikeType,
-         dim: Optional[str] = None,
-         *,
-         out: Optional[VariableLikeType] = None) -> VariableLikeType:
+def mean(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Arithmetic mean of elements in the input.
 
     If the input has variances, the variances stored in the output are based on
@@ -32,8 +28,6 @@ def mean(x: VariableLikeType,
     dim:
         Dimension along which to calculate the mean. If not
         given, the mean over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -46,15 +40,12 @@ def mean(x: VariableLikeType,
         Ignore NaN's when calculating the mean.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.mean, x, out=out)
+        return _cpp.mean(x)
     else:
-        return _call_cpp_func(_cpp.mean, x, dim=dim, out=out)
+        return _cpp.mean(x, dim=dim)
 
 
-def nanmean(x: VariableLikeType,
-            dim: Optional[str] = None,
-            *,
-            out: Optional[VariableLikeType] = None) -> VariableLikeType:
+def nanmean(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Arithmetic mean of elements in the input ignoring NaN's.
 
     If the input has variances, the variances stored in the output are based on
@@ -73,8 +64,6 @@ def nanmean(x: VariableLikeType,
     dim:
         Dimension along which to calculate the mean. If not
         given, the nanmean over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -87,15 +76,12 @@ def nanmean(x: VariableLikeType,
         Compute the mean without special handling of NaN.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.nanmean, x, out=out)
+        return _cpp.nanmean(x)
     else:
-        return _call_cpp_func(_cpp.nanmean, x, dim=dim, out=out)
+        return _cpp.nanmean(x, dim=dim)
 
 
-def sum(x: VariableLikeType,
-        dim: Optional[str] = None,
-        *,
-        out: Optional[VariableLikeType] = None) -> VariableLikeType:
+def sum(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Sum of elements in the input.
 
     If the input data is in single precision (dtype='float32') this internally uses
@@ -111,8 +97,6 @@ def sum(x: VariableLikeType,
     dim:
         Optional dimension along which to calculate the sum. If not
         given, the sum over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -125,15 +109,12 @@ def sum(x: VariableLikeType,
         Ignore NaN's when calculating the sum.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.sum, x, out=out)
+        return _cpp.sum(x)
     else:
-        return _call_cpp_func(_cpp.sum, x, dim=dim, out=out)
+        return _cpp.sum(x, dim=dim)
 
 
-def nansum(x: VariableLikeType,
-           dim: Optional[str] = None,
-           *,
-           out: Optional[VariableLikeType] = None) -> VariableLikeType:
+def nansum(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Sum of elements in the input ignoring NaN's.
 
     See :py:func:`scipp.sum` on how rounding errors for float32 inputs are handled.
@@ -145,8 +126,6 @@ def nansum(x: VariableLikeType,
     dim:
         Optional dimension along which to calculate the sum. If not
         given, the sum over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -159,26 +138,21 @@ def nansum(x: VariableLikeType,
        Compute the sum without special handling of NaN.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.nansum, x, out=out)
+        return _cpp.nansum(x)
     else:
-        return _call_cpp_func(_cpp.nansum, x, dim=dim, out=out)
+        return _cpp.nansum(x, dim=dim)
 
 
-def min(x: _cpp.Variable,
-        dim: Optional[str] = None,
-        *,
-        out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def min(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Minimum of elements in the input.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the min. If not
         given, the min over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -195,26 +169,21 @@ def min(x: _cpp.Variable,
         Same as max but ignoring NaN's.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.min, x, out=out)
+        return _cpp.min(x)
     else:
-        return _call_cpp_func(_cpp.min, x, dim=dim, out=out)
+        return _cpp.min(x, dim=dim)
 
 
-def max(x: _cpp.Variable,
-        dim: Optional[str] = None,
-        *,
-        out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def max(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Maximum of elements in the input.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the max. If not
         given, the max over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -231,26 +200,21 @@ def max(x: _cpp.Variable,
         Same as max but ignoring NaN's.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.max, x, out=out)
+        return _cpp.max(x)
     else:
-        return _call_cpp_func(_cpp.max, x, dim=dim, out=out)
+        return _cpp.max(x, dim=dim)
 
 
-def nanmin(x: _cpp.Variable,
-           dim: Optional[str] = None,
-           *,
-           out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def nanmin(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Minimum of elements in the input ignoring NaN's.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the min. If not
         given, the min over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -267,26 +231,21 @@ def nanmin(x: _cpp.Variable,
         Same as max but ignoring NaN's.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.nanmin, x, out=out)
+        return _cpp.nanmin(x)
     else:
-        return _call_cpp_func(_cpp.nanmin, x, dim=dim, out=out)
+        return _cpp.nanmin(x, dim=dim)
 
 
-def nanmax(x: _cpp.Variable,
-           dim: Optional[str] = None,
-           *,
-           out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def nanmax(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Maximum of elements in the input ignoring NaN's.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the max. If not
         given, the max over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -303,26 +262,21 @@ def nanmax(x: _cpp.Variable,
         Same as min but ignoring NaN's.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.nanmax, x, out=out)
+        return _cpp.nanmax(x)
     else:
-        return _call_cpp_func(_cpp.nanmax, x, dim=dim, out=out)
+        return _cpp.nanmax(x, dim=dim)
 
 
-def all(x: _cpp.Variable,
-        dim: Optional[str] = None,
-        *,
-        out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def all(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Logical AND over input values.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the AND. If not
         given, the AND over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -336,26 +290,21 @@ def all(x: _cpp.Variable,
         Logical OR.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.all, x, out=out)
+        return _cpp.all(x)
     else:
-        return _call_cpp_func(_cpp.all, x, dim=dim, out=out)
+        return _cpp.all(x, dim=dim)
 
 
-def any(x: _cpp.Variable,
-        dim: Optional[str] = None,
-        *,
-        out: Optional[_cpp.Variable] = None) -> _cpp.Variable:
+def any(x: VariableLikeType, dim: Optional[str] = None) -> VariableLikeType:
     """Logical OR over input values.
 
     Parameters
     ----------
-    x:
+    x: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the OR. If not
         given, the OR over all dimensions is calculated.
-    out:
-        Optional output buffer.
 
     Returns
     -------
@@ -369,6 +318,6 @@ def any(x: _cpp.Variable,
         Logical AND.
     """
     if dim is None:
-        return _call_cpp_func(_cpp.any, x, out=out)
+        return _cpp.any(x)
     else:
-        return _call_cpp_func(_cpp.any, x, dim=dim, out=out)
+        return _cpp.any(x, dim=dim)