+++ src/scipp/core/variable.py
@@ -59,7 +59,7 @@ def scalar(value: Any,
       <scipp.Variable> ()    float64  [dimensionless]  [3.14]
 
       >>> sc.scalar('a string')
-      <scipp.Variable> ()     string           [None]  ["a string"]
+      <scipp.Variable> ()     string           <no unit>  ["a string"]
 
     Or specifying a unit and dtype:
 
@@ -70,7 +70,7 @@ def scalar(value: Any,
     object in a scalar variable and *not* create an array variable:
 
       >>> sc.scalar([1, 2, 3])
-      <scipp.Variable> ()   PyObject           [None]  [[1, 2, 3]]
+      <scipp.Variable> ()   PyObject           <no unit>  [[1, 2, 3]]
     """
     return _cpp.Variable(dims=(),
                          values=value,
@@ -105,7 +105,7 @@ def index(value: Any, *, dtype: Optional[DTypeLike] = None) -> Variable:
     --------
 
       >>> sc.index(123)
-      <scipp.Variable> ()      int64           [None]  [123]
+      <scipp.Variable> ()      int64           <no unit>  [123]
     """
     return scalar(value=value, dtype=dtype, unit=None)
 
@@ -808,7 +808,10 @@ def datetime(value: Union[str, int, _np.datetime64],
       <scipp.Variable> ()  datetime64              [s]  [2021-01-10T14:16:15]
     """
     if isinstance(value, str):
-        return scalar(_np.datetime64(value), unit=unit)
+        try:
+            return scalar(_np.datetime64(value), unit=unit)
+        except DeprecationWarning as e:
+            raise ValueError('Parsing timezone aware datetimes is not supported')
     return scalar(value, unit=unit, dtype=_cpp.DType.datetime64)
 
 
@@ -850,8 +853,11 @@ def datetimes(*,
         np_unit_str = ''
     else:
         np_unit_str = f'[{_cpp.to_numpy_time_string(unit)}]'
-    return array(dims=dims,
-                 values=_np.asarray(values, dtype=f'datetime64{np_unit_str}'))
+    try:
+        return array(dims=dims,
+                     values=_np.asarray(values, dtype=f'datetime64{np_unit_str}'))
+    except DeprecationWarning as e:
+        raise ValueError('Parsing timezone aware datetimes is not supported')
 
 
 def epoch(*, unit: Union[Unit, str]) -> Variable: